Moving code to util.c
